Ajuste a variable long

+++ src/main/java/com/curso/proyecto/controllers/LibrosController.java
@@ -5,7 +5,7 @@ import java.util.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
+//import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -21,7 +21,7 @@ public class LibrosController {
 
     // Crear libro
     @PostMapping("/api/libro")
-    public Integer createLibro(@RequestBody LibrosModel libro) {
+    public Long createLibro(@RequestBody LibrosModel libro) {
         return librosService.createLibro(libro);
     }
