+++ src/main/java/com/curso/proyecto/repositories/LibrosRepository.java
@@ -3,6 +3,6 @@ package com.curso.proyecto.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.curso.proyecto.models.LibrosModel;
 
-public interface LibrosRepository extends JpaRepository<LibrosModel, Integer>{
+public interface LibrosRepository extends JpaRepository<LibrosModel, Long>{
     
 }
